Add tests, assert no operations after TxMgr is closed.
