+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/ReadOnlyTransactionManager.java
@@ -15,6 +15,7 @@
  */
 package com.palantir.atlasdb.transaction.impl;
 
+import com.google.common.base.Preconditions;
 import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 import com.palantir.atlasdb.keyvalue.api.Cell;
@@ -76,6 +77,7 @@ public class ReadOnlyTransactionManager extends AbstractTransactionManager imple
         this.startTimestamp = startTimestamp;
         this.readSentinelBehavior = readSentinelBehavior;
         this.allowHiddenTableAccess = allowHiddenTableAccess;
+        this.isClosed = false;
     }
 
     @Override
@@ -108,6 +110,7 @@ public class ReadOnlyTransactionManager extends AbstractTransactionManager imple
 
     @Override
     public void close() {
+        super.close();
         keyValueService.close();
     }
