+++ atlasdb-tests-shared/src/main/java/com/palantir/atlasdb/transaction/impl/AbstractSerializableTransactionTest.java
@@ -63,7 +63,7 @@ import com.palantir.lock.LockRefreshToken;
 public abstract class AbstractSerializableTransactionTest extends AbstractTransactionTest {
 
     @Override
-    protected TransactionManager getManager() {
+    protected TransactionManager createManager() {
         return new SerializableTransactionManager(
                 keyValueService,
                 timestampService,
